% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_body.R
\name{generate_body}
\alias{generate_body}
\title{Generate HTML body}
\usage{
generate_body(df, max_rows = 20, drop = FALSE)
}
\arguments{
\item{df}{data frame}

\item{max_rows}{integer}

\item{drop}{boolean}
}
\value{
string
}
\description{
Generate HTML body
}
