% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebook.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Generate codebook}
\usage{
generate_codebook(df, max_rows = 20, drop = FALSE)
}
\arguments{
\item{df}{data frame}

\item{max_rows}{integer}

\item{drop}{boolean}
}
\value{
string
}
\description{
Generate codebook
}
\examples{
\dontrun{
    reader::write_lines(generate_codebook(mtcars), 'mtcars codebook.html')
}
}
