% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_left_pane.R
\name{generate_left_pane}
\alias{generate_left_pane}
\title{Generate navigation pane of the codebook}
\usage{
generate_left_pane(var_names)
}
\arguments{
\item{var_names}{character vector}
}
\value{
string
}
\description{
Generate navigation pane of the codebook
}
