% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_list_entry.R
\name{generate_list_entry}
\alias{generate_list_entry}
\title{Generate navigation item for variable}
\usage{
generate_list_entry(var)
}
\arguments{
\item{var}{variable name}
}
\value{
string
}
\description{
Generate navigation item for variable
}
