% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_main_content.R
\name{generate_main_content}
\alias{generate_main_content}
\title{Generate main content with variable entries}
\usage{
generate_main_content(df, max_rows = 20, drop = FALSE)
}
\arguments{
\item{df}{data frame}

\item{max_rows}{integer}

\item{drop}{boolean}
}
\value{
string
}
\description{
Generate main content with variable entries
}
