% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_entry.R
\name{generate_var_entry}
\alias{generate_var_entry}
\title{Generate variable entry in main content}
\usage{
generate_var_entry(x, var, max_rows = 20, drop = FALSE)
}
\arguments{
\item{x}{vector}

\item{max_rows}{integer}

\item{drop}{boolean}

\item{text}{string}
}
\value{
string
}
\description{
Generate variable entry in main content
}
