% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_label.R
\name{generate_var_label}
\alias{generate_var_label}
\title{Generate variable label}
\usage{
generate_var_label(x)
}
\arguments{
\item{x}{vector}
}
\value{
string
}
\description{
Generate variable label
}
