% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_name.R
\name{generate_var_name}
\alias{generate_var_name}
\title{Create HTML element for variable name}
\usage{
generate_var_name(x)
}
\arguments{
\item{x}{vector}
}
\value{
character vector
}
\description{
Create HTML element for variable name
}
