% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_table.R
\name{generate_var_table}
\alias{generate_var_table}
\title{Generate Value-Label table}
\usage{
generate_var_table(x, max_rows, drop = FALSE)
}
\arguments{
\item{x}{vector}

\item{max_rows}{integer}

\item{drop}{boolean}
}
\value{
string
}
\description{
Generate Value-Label table
}
