% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_text.R
\name{generate_var_text}
\alias{generate_var_text}
\title{Generate variable text/note}
\usage{
generate_var_text(x)
}
\arguments{
\item{x}{vector}
}
\value{
character vector
}
\description{
Generate variable text/note
}
