% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_surround.R
\name{tag_surround}
\alias{tag_surround}
\title{Surround entry with HTML tag}
\usage{
tag_surround(x, tag, ...)
}
\arguments{
\item{x}{content of tag}

\item{tag}{tag name}

\item{...}{attributes for tag}
}
\value{
character vector
}
\description{
Surround entry with HTML tag
}
